Moved some utility functions to a test-specific utility compilation-unit
